Revert "feat: Speed up sequence diagram gen"
This reverts commit fa9c8984c0a641822d8ac1abf4a529f958bd2a35.

Archiving AppMaps from /home/runner/work/rails_tutorial_sample_app_7th_ed/rails_tutorial_sample_app_7th_ed
Warning: Command "appmap archive --revision da020755ab1dd8a56fab03563f59c1ae85e7b14e" exited with failure code 1
Warning: Error: Cannot find module '/snapshot/appmap-js/node_modules/piscina/dist/src/worker.js'
    at Function.Module._resolveFilename (node:internal/modules/cjs/loader:933:15)
    at Function._resolveFilename (pkg/prelude/bootstrap.js:1955:46)
    at Function.Module._load (node:internal/modules/cjs/loader:778:27)
    at Function.executeUserEntryPoint [as runMain] (node:internal/modules/run_main:77:12)
    at MessagePort.<anonymous> (node:internal/main/worker_thread:191:24)
    at MessagePort.[nodejs.internal.kHybridDispatch] (node:internal/event_target:643:20)
    at MessagePort.exports.emitMessage (node:internal/per_context/messageport:23:28) {
  code: 'MODULE_NOT_FOUND',
  requireStack: []
}
