+++ packages/cli/src/cmds/archive/updateSequenceDiagrams.ts
@@ -1,12 +1,22 @@
-import Piscina from 'piscina';
+import { AppMapFilter, buildAppMap } from '@appland/models';
+import {
+  buildDiagram,
+  format,
+  FormatType,
+  SequenceDiagramOptions,
+  Specification,
+} from '@appland/sequence-diagram';
+import { queue } from 'async';
+import { readFile, stat, unlink, writeFile } from 'fs/promises';
+import { glob } from 'glob';
+import { basename, dirname, join } from 'path';
 import { promisify } from 'util';
-import glob from 'glob';
-import { cpus } from 'os';
-import { stat } from 'fs/promises';
-import { join, resolve } from 'path';
-import { AppMapFilter, serializeFilter } from '@appland/models';
-import { SequenceDiagramOptions } from '@appland/sequence-diagram';
+import { processFiles } from '../../utils';
+import FileTooLargeError from '../../fingerprint/fileTooLargeError';
+import { CountNumProcessed } from './CountNumProcessed';
 import reportAppMapProcessingError from './reportAppMapProcessingError';
+
+const Concurrency = 5;
 
 export default async function updateSequenceDiagrams(
   dir: string,
@@ -19,40 +29,32 @@ export default async function updateSequenceDiagrams(
 
   const oversizedAppMaps = new Array<string>();
 
-  const cores = Math.round(cpus().length / 4);
-  console.log(`Using ${cores}..${cores * 2} cores to generate sequence diagrams`);
-  const piscina = new Piscina({
-    filename: resolve(__dirname, 'workers/buildSequenceDiagram.js'),
-    minThreads: cores,
-    maxThreads: cores * 2,
-  });
-  piscina.on('error', reportAppMapProcessingError('Sequence diagram'));
-
-  const files = await promisify(glob)(join(dir, '**', '*.appmap.json'));
-  const filesToProcess = new Array<string>();
-  for (const file of files) {
-    const stats = await stat(file);
+  const generateDiagram = async (appmapFileName: string) => {
+    // Determine size of file appmapFileName in bytes
+    const stats = await stat(appmapFileName);
     if (stats.size > maxAppMapSizeInBytes) {
-      oversizedAppMaps.push(file);
-      continue;
+      throw new FileTooLargeError(appmapFileName, stats.size, maxAppMapSizeInBytes);
     }
 
-    filesToProcess.push(file);
-  }
+    const fullAppMap = buildAppMap()
+      .source(await readFile(appmapFileName, 'utf8'))
+      .build();
+    const filteredAppMap = filter.filter(fullAppMap, []);
+    const specification = Specification.build(filteredAppMap, specOptions);
+    const diagram = buildDiagram(appmapFileName, filteredAppMap, specification);
+    const diagramOutput = format(FormatType.JSON, diagram, appmapFileName);
+    const indexDir = join(dirname(appmapFileName), basename(appmapFileName, '.appmap.json'));
+    const diagramFileName = join(indexDir, 'sequence.json');
+    await writeFile(diagramFileName, diagramOutput.diagram);
+  };
 
-  const filterStr = serializeFilter(filter);
-  const startTime = new Date().getTime();
-  await Promise.all(
-    filesToProcess.map((file) => piscina.run({ appmapFileName: file, filterStr, specOptions }))
+  const counter = new CountNumProcessed();
+  await processFiles(
+    join(dir, '**', '*.appmap.json'),
+    generateDiagram,
+    counter.setCount(),
+    reportAppMapProcessingError('Sequence diagram')
   );
-  const elapsed = new Date().getTime() - startTime;
-  const sequenceDiagrams = await promisify(glob)(join(dir, '**', 'sequence.json'));
-  console.log(`Generated ${sequenceDiagrams.length} sequence digrams in ${elapsed}ms`);
-  if (sequenceDiagrams.length !== filesToProcess.length) {
-    console.warn(
-      `Expected ${filesToProcess.length} sequence diagrams, but found ${sequenceDiagrams.length}`
-    );
-  }
 
-  return { numGenerated: files.length, oversizedAppMaps };
+  return { numGenerated: counter.count, oversizedAppMaps };
 }
